fix 64-bit indexing problem

+++ cpp/mountainsort/src/processors/extract_clips.cpp
@@ -8,20 +8,20 @@
 
 Mda extract_clips(const DiskReadMda& X, const QVector<double>& times, int clip_size)
 {
-    int M = X.N1();
-    int N = X.N2();
-    int T = clip_size;
-    int L = times.count();
-    int Tmid = (int)((T + 1) / 2) - 1;
+    bigint M = X.N1();
+    bigint N = X.N2();
+    bigint T = clip_size;
+    bigint L = times.count();
+    bigint Tmid = (bigint)((T + 1) / 2) - 1;
     Mda clips(M, T, L);
-    for (int i = 0; i < L; i++) {
-        int t1 = (int)times[i] - Tmid;
-        int t2 = t1 + T - 1;
+    for (bigint i = 0; i < L; i++) {
+        bigint t1 = (bigint)times[i] - Tmid;
+        bigint t2 = t1 + T - 1;
         if ((t1 >= 0) && (t2 < N)) {
             Mda tmp;
             X.readChunk(tmp, 0, t1, M, T);
-            for (int t = 0; t < T; t++) {
-                for (int m = 0; m < M; m++) {
+            for (bigint t = 0; t < T; t++) {
+                for (bigint m = 0; m < M; m++) {
                     clips.set(tmp.get(m, t), m, t, i);
                 }
             }
@@ -32,20 +32,20 @@ Mda extract_clips(const DiskReadMda& X, const QVector<double>& times, int clip_s
 
 Mda32 extract_clips(const DiskReadMda32& X, const QVector<double>& times, int clip_size)
 {
-    int M = X.N1();
-    int N = X.N2();
-    int T = clip_size;
-    int L = times.count();
-    int Tmid = (int)((T + 1) / 2) - 1;
+    bigint M = X.N1();
+    bigint N = X.N2();
+    bigint T = clip_size;
+    bigint L = times.count();
+    bigint Tmid = (bigint)((T + 1) / 2) - 1;
     Mda32 clips(M, T, L);
-    for (int i = 0; i < L; i++) {
-        int t1 = (int)times[i] - Tmid;
-        int t2 = t1 + T - 1;
+    for (bigint i = 0; i < L; i++) {
+        bigint t1 = (bigint)times[i] - Tmid;
+        bigint t2 = t1 + T - 1;
         if ((t1 >= 0) && (t2 < N)) {
             Mda32 tmp;
             X.readChunk(tmp, 0, t1, M, T);
-            for (int t = 0; t < T; t++) {
-                for (int m = 0; m < M; m++) {
+            for (bigint t = 0; t < T; t++) {
+                for (bigint m = 0; m < M; m++) {
                     clips.set(tmp.get(m, t), m, t, i);
                 }
             }
@@ -56,21 +56,21 @@ Mda32 extract_clips(const DiskReadMda32& X, const QVector<double>& times, int cl
 
 Mda extract_clips(const DiskReadMda& X, const QVector<double>& times, const QVector<int>& channels, int clip_size)
 {
-    int M = X.N1();
-    int N = X.N2();
-    int M0 = channels.count();
-    int T = clip_size;
-    int L = times.count();
-    int Tmid = (int)((T + 1) / 2) - 1;
+    bigint M = X.N1();
+    bigint N = X.N2();
+    bigint M0 = channels.count();
+    bigint T = clip_size;
+    bigint L = times.count();
+    bigint Tmid = (bigint)((T + 1) / 2) - 1;
     Mda clips(M0, T, L);
-    for (int i = 0; i < L; i++) {
-        int t1 = (int)times[i] - Tmid;
-        int t2 = t1 + T - 1;
+    for (bigint i = 0; i < L; i++) {
+        bigint t1 = (bigint)times[i] - Tmid;
+        bigint t2 = t1 + T - 1;
         if ((t1 >= 0) && (t2 < N)) {
             Mda tmp;
             X.readChunk(tmp, 0, t1, M, T);
-            for (int t = 0; t < T; t++) {
-                for (int m0 = 0; m0 < M0; m0++) {
+            for (bigint t = 0; t < T; t++) {
+                for (bigint m0 = 0; m0 < M0; m0++) {
                     clips.set(tmp.get(channels[m0], t), m0, t, i);
                 }
             }
@@ -81,21 +81,21 @@ Mda extract_clips(const DiskReadMda& X, const QVector<double>& times, const QVec
 
 Mda32 extract_clips(const DiskReadMda32& X, const QVector<double>& times, const QVector<int>& channels, int clip_size)
 {
-    int M = X.N1();
-    int N = X.N2();
-    int M0 = channels.count();
-    int T = clip_size;
-    int L = times.count();
-    int Tmid = (int)((T + 1) / 2) - 1;
+    bigint M = X.N1();
+    bigint N = X.N2();
+    bigint M0 = channels.count();
+    bigint T = clip_size;
+    bigint L = times.count();
+    bigint Tmid = (bigint)((T + 1) / 2) - 1;
     Mda32 clips(M0, T, L);
-    for (int i = 0; i < L; i++) {
-        int t1 = (int)times[i] - Tmid;
-        int t2 = t1 + T - 1;
+    for (bigint i = 0; i < L; i++) {
+        bigint t1 = (bigint)times[i] - Tmid;
+        bigint t2 = t1 + T - 1;
         if ((t1 >= 0) && (t2 < N)) {
             Mda32 tmp;
             X.readChunk(tmp, 0, t1, M, T);
-            for (int t = 0; t < T; t++) {
-                for (int m0 = 0; m0 < M0; m0++) {
+            for (bigint t = 0; t < T; t++) {
+                for (bigint m0 = 0; m0 < M0; m0++) {
                     clips.set(tmp.get(channels[m0], t), m0, t, i);
                 }
             }
@@ -107,19 +107,19 @@ Mda32 extract_clips(const DiskReadMda32& X, const QVector<double>& times, const 
 bool extract_clips(const QString& timeseries_path, const QString& firings_path, const QString& clips_path, int clip_size, const QList<int>& channels_in, double t1, double t2)
 {
     QVector<int> channels;
-    for (int i = 0; i < channels_in.count(); i++)
+    for (bigint i = 0; i < channels_in.count(); i++)
         channels << channels_in[i] - 1;
     DiskReadMda X(timeseries_path);
     DiskReadMda F(firings_path);
 
     if (channels.isEmpty()) {
-        for (int m = 0; m < X.N1(); m++) {
+        for (bigint m = 0; m < X.N1(); m++) {
             channels << m;
         }
     }
 
     QVector<double> times;
-    for (int j = 0; j < F.N2(); j++) {
+    for (bigint j = 0; j < F.N2(); j++) {
         double t0 = F.value(1, j);
         if ((t2 == 0) || ((t1 <= t0) && (t0 <= t2)))
             times << F.value(1, j);
